%$HeadURL$
\encoding{utf8}
\name{sprof-package}
\alias{sprof-package}
\alias{sprof}
\docType{package}
\title{sprof: Analysis of R profiles}
\description{
Profiling, timing and optimization utilitites}
\details{
\tabular{ll}{
Package: \tab sprof\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2013-05-26\cr
License: \tab GPL-2\cr
}
\code{readProf} reads a profile file from \code{Rprof} or other profilers and returns a composite structure of class \code{sprof}.
}
\author{
Günther Sawitzki <gsawitzki@r-forge.r-project.org>}
%\references{
%~~ Literature or other references for background information ~~
%}
\keyword{utilities}
\seealso{
\code{\link[proftools:proftools-package]{proftools}}.
 
\code{\link[profr:profr-package]{profr}}.

\code{\link{profr}}.
 }
%\examples{
%~~ simple examples of the most important functions ~~
%}
